requirements.txt
APScheduler==3.8.1
asgiref==3.5.0
autopep8==1.6.0
Deprecated==1.2.13
Django==4.0.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
wrapt==1.13.3
zipp==3.7.0
